+++ src/CntkCatalyst/MemoryMinibatchSource.cs
@@ -63,9 +63,9 @@ namespace CntkCatalyst
         public string FeaturesName => m_featuresName;
         public string TargetsName => m_targetsName;
 
-        public IDictionary<StreamInformation, MinibatchData> GetNextMinibatch(uint minibatchSizeInSamples, DeviceDescriptor device)
+        public IDictionary<StreamInformation, MinibatchData> GetNextMinibatch(int minibatchSizeInSamples, DeviceDescriptor device)
         {
-            var minibatchData = GetNextMinibatch((int)minibatchSizeInSamples);
+            var minibatchData = GetNextMinibatch(minibatchSizeInSamples);
 
             var batchObservations = Value.CreateBatch<float>(m_observations.SampleShape, minibatchData.observations, device, true);
             var batchTarget = Value.CreateBatch<float>(m_targets.SampleShape, minibatchData.targets, device, true);
@@ -73,10 +73,10 @@ namespace CntkCatalyst
             var minibatch = new Dictionary<StreamInformation, MinibatchData>
             {
                 { StreamInfo(m_featuresName), new MinibatchData(batchObservations,
-                    minibatchSizeInSamples, minibatchData.isSweepEnd) },
+                    (uint)minibatchSizeInSamples, minibatchData.isSweepEnd) },
 
                 { StreamInfo(m_targetsName), new MinibatchData(batchTarget,
-                    minibatchSizeInSamples, minibatchData.isSweepEnd) },
+                    (uint)minibatchSizeInSamples, minibatchData.isSweepEnd) },
             };
 
             return minibatch;
